requirements.txt
streamlit
pandas
openpyxl
pdfplumber
pydantic
python-dotenv
openai
